LayoutTests/http/tests/websocket/tests/hybi/handshake-fail-by-more-accept-header-expected.txt

CONSOLE MESSAGE: The Sec-WebSocket-Accept header MUST NOT appear more than once in an HTTP response
Test that WebSocket handshake fails if there are more one Sec-WebSocket-Accept header field in the response.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE
